requirements/local.txt
-r base.txt
flake8==7.0.0
black==24.4.2
isort==5.13.2
pytest-cov==5.0.0
pytest-django==4.8.0
Faker==0.7.4
pytest-factoryboy==2.7.0
